+++ src/components.d.ts
@@ -5,7 +5,7 @@
  * It contains typing information for all components that exist in this project.
  */
 import { HTMLStencilElement, JSXBase } from "@stencil/core/internal";
-import { ButtonSizeEnum, ButtonVariantBean, ComboItemBean, DividerOrientation, DividerSize, HeaderUserData, InputStatusBean, InputTypeBean, MenuItem, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, TableHeaderSize, TabOrientationBean, TabSizeBean, ThemeVariantBean } from "./beans";
+import { ButtonSizeEnum, ButtonVariantBean, ComboItemBean, DividerOrientation, DividerSize, HeaderUserData, InputStatusBean, InputTypeBean, MenuItem, PopoverBorderRadius, PopoverPosition, PopoverShadow, SelectItemBean, StatusTagStatus, TableHeaderSize, ThemeVariant, ThemeVariantBean } from "./beans";
 import { ListItemBean } from "./beans/index.js";
 import { LicenseTypeEnum, MenuItem as MenuItem1, TooltipPosition } from "./beans/index";
 import { ZTypographyLevels } from "./components/typography/z-typography/index";
@@ -988,6 +988,28 @@ export namespace Components {
          */
         "status"?: InputStatusBean;
     }
+    interface ZStatusTag {
+        /**
+          * [optional] Hide the text and show it on hover
+         */
+        "expandable"?: boolean;
+        /**
+          * [optional] Status tag icon
+         */
+        "icon"?: string;
+        /**
+          * [optional] Status tag color
+         */
+        "status"?: StatusTagStatus;
+        /**
+          * [optional] Status tag text
+         */
+        "text"?: string;
+        /**
+          * [optional] Status tag style
+         */
+        "variant"?: ThemeVariant;
+    }
     interface ZStepper {
     }
     interface ZStepperItem {
@@ -1393,6 +1415,12 @@ declare global {
         prototype: HTMLZSelectElement;
         new (): HTMLZSelectElement;
     };
+    interface HTMLZStatusTagElement extends Components.ZStatusTag, HTMLStencilElement {
+    }
+    var HTMLZStatusTagElement: {
+        prototype: HTMLZStatusTagElement;
+        new (): HTMLZStatusTagElement;
+    };
     interface HTMLZStepperElement extends Components.ZStepper, HTMLStencilElement {
     }
     var HTMLZStepperElement: {
@@ -1486,6 +1514,7 @@ declare global {
         "z-registro-table-row": HTMLZRegistroTableRowElement;
         "z-registro-table-sticky-footer": HTMLZRegistroTableStickyFooterElement;
         "z-select": HTMLZSelectElement;
+        "z-status-tag": HTMLZStatusTagElement;
         "z-stepper": HTMLZStepperElement;
         "z-stepper-item": HTMLZStepperItemElement;
         "z-toggle-button": HTMLZToggleButtonElement;
@@ -2566,6 +2595,28 @@ declare namespace LocalJSX {
          */
         "status"?: InputStatusBean;
     }
+    interface ZStatusTag {
+        /**
+          * [optional] Hide the text and show it on hover
+         */
+        "expandable"?: boolean;
+        /**
+          * [optional] Status tag icon
+         */
+        "icon"?: string;
+        /**
+          * [optional] Status tag color
+         */
+        "status"?: StatusTagStatus;
+        /**
+          * [optional] Status tag text
+         */
+        "text"?: string;
+        /**
+          * [optional] Status tag style
+         */
+        "variant"?: ThemeVariant;
+    }
     interface ZStepper {
     }
     interface ZStepperItem {
@@ -2702,6 +2753,7 @@ declare namespace LocalJSX {
         "z-registro-table-row": ZRegistroTableRow;
         "z-registro-table-sticky-footer": ZRegistroTableStickyFooter;
         "z-select": ZSelect;
+        "z-status-tag": ZStatusTag;
         "z-stepper": ZStepper;
         "z-stepper-item": ZStepperItem;
         "z-toggle-button": ZToggleButton;
@@ -2770,6 +2822,7 @@ declare module "@stencil/core" {
             "z-registro-table-row": LocalJSX.ZRegistroTableRow & JSXBase.HTMLAttributes<HTMLZRegistroTableRowElement>;
             "z-registro-table-sticky-footer": LocalJSX.ZRegistroTableStickyFooter & JSXBase.HTMLAttributes<HTMLZRegistroTableStickyFooterElement>;
             "z-select": LocalJSX.ZSelect & JSXBase.HTMLAttributes<HTMLZSelectElement>;
+            "z-status-tag": LocalJSX.ZStatusTag & JSXBase.HTMLAttributes<HTMLZStatusTagElement>;
             "z-stepper": LocalJSX.ZStepper & JSXBase.HTMLAttributes<HTMLZStepperElement>;
             "z-stepper-item": LocalJSX.ZStepperItem & JSXBase.HTMLAttributes<HTMLZStepperItemElement>;
             "z-toggle-button": LocalJSX.ZToggleButton & JSXBase.HTMLAttributes<HTMLZToggleButtonElement>;